backend/requirements.txt
click==8.1.3
Flask==2.1.3
Flask-Cors==3.0.10
gunicorn==20.1.0
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
psycopg2-binary==2.9.3
six==1.16.0
Werkzeug==2.2.1
zipp==3.8.1
